% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tot.R
\name{combine_tot}
\alias{combine_tot}
\title{Combine y by x}
\usage{
combine_tot(x)
}
\arguments{
\item{x}{A 2 col dataframe}
}
\description{
Combine x by x
}
\examples{
x <- data.frame(a= c('a', 'a', 's', 'f', 'd', 'a', 's', 's'), b = state.name[1:8])
combine_tot(x)

#DESIRED OUTPUT
##    a                    b
## 1: a       Alabama Alaska
## 2: s              Arizona
## 3: f             Arkansas
## 4: d           California
## 5: a             Colorado
## 6: s Connecticut Delaware
}
